+++ sdk/storage/azure-storage-blob/src/main/java/com/azure/storage/blob/models/BlobProperties.java
@@ -43,9 +43,11 @@ public final class BlobProperties {
     private final Boolean isAccessTierInferred;
     private final ArchiveStatus archiveStatus;
     private final String encryptionKeySha256;
+    private final String encryptionScope;
     private final OffsetDateTime accessTierChangeTime;
     private final Map<String, String> metadata;
     private final Integer committedBlockCount;
+    private final Long tagCount;
     private final String versionId;
     private final Boolean isCurrentVersion;
     private final Map<String, ObjectReplicationPolicy> objectReplicationSourcePolicies;
@@ -102,10 +104,11 @@ public final class BlobProperties {
         final OffsetDateTime accessTierChangeTime, final Map<String, String> metadata,
         final Integer committedBlockCount) {
         this(creationTime, lastModified, eTag, blobSize, contentType, contentMd5, contentEncoding, contentDisposition,
-             contentLanguage, cacheControl, blobSequenceNumber, blobType, leaseStatus, leaseState, leaseDuration,
+            contentLanguage, cacheControl, blobSequenceNumber, blobType, leaseStatus, leaseState, leaseDuration,
             copyId, copyStatus, copySource, copyProgress, copyCompletionTime, copyStatusDescription, isServerEncrypted,
             isIncrementalCopy, copyDestinationSnapshot, accessTier, isAccessTierInferred, archiveStatus,
-            encryptionKeySha256, accessTierChangeTime, metadata, committedBlockCount, null, null, null);
+            encryptionKeySha256, null, accessTierChangeTime, metadata, committedBlockCount, null, null,
+            null, null, (String) null);
     }
 
     /**
@@ -141,73 +144,14 @@ public final class BlobProperties {
      * blob.
      * @param archiveStatus Archive status of the blob.
      * @param encryptionKeySha256 SHA256 of the customer provided encryption key used to encrypt the blob on the server.
+     * @param encryptionScope The name of the encryption scope under which the blob is encrypted.
      * @param accessTierChangeTime Datetime when the access tier of the blob last changed.
      * @param metadata Metadata associated with the blob.
      * @param committedBlockCount Number of blocks committed to an append blob, if the blob is a block or page blob
      * pass {@code null}.
      * @param versionId The version identifier of the blob.
      * @param isCurrentVersion Flag indicating if version identifier points to current version of the blob.
-     * @param objectReplicationStatus The object replication status map to parse.
-     */
-    public BlobProperties(final OffsetDateTime creationTime, final OffsetDateTime lastModified, final String eTag,
-        final long blobSize, final String contentType, final byte[] contentMd5, final String contentEncoding,
-        final String contentDisposition, final String contentLanguage, final String cacheControl,
-        final Long blobSequenceNumber, final BlobType blobType, final LeaseStatusType leaseStatus,
-        final LeaseStateType leaseState, final LeaseDurationType leaseDuration, final String copyId,
-        final CopyStatusType copyStatus, final String copySource, final String copyProgress,
-        final OffsetDateTime copyCompletionTime, final String copyStatusDescription, final Boolean isServerEncrypted,
-        final Boolean isIncrementalCopy, final String copyDestinationSnapshot, final AccessTier accessTier,
-        final Boolean isAccessTierInferred, final ArchiveStatus archiveStatus, final String encryptionKeySha256,
-        final OffsetDateTime accessTierChangeTime, final Map<String, String> metadata,
-        final Integer committedBlockCount, final String versionId, final Boolean isCurrentVersion,
-        Map<String, String> objectReplicationStatus) {
-        this(creationTime, lastModified, eTag, blobSize, contentType, contentMd5, contentEncoding, contentDisposition,
-            contentLanguage, cacheControl, blobSequenceNumber, blobType, leaseStatus, leaseState, leaseDuration, copyId,
-            copyStatus, copySource, copyProgress, copyCompletionTime, copyStatusDescription, isServerEncrypted,
-            isIncrementalCopy, copyDestinationSnapshot, accessTier, isAccessTierInferred, archiveStatus,
-            encryptionKeySha256, accessTierChangeTime, metadata, committedBlockCount, versionId, isCurrentVersion,
-            objectReplicationStatus, null);
-    }
-
-    /**
-     * Constructs a {@link BlobProperties}.
-     *
-     * @param creationTime Creation time of the blob.
-     * @param lastModified Datetime when the blob was last modified.
-     * @param eTag ETag of the blob.
-     * @param blobSize Size of the blob.
-     * @param contentType Content type specified for the blob.
-     * @param contentMd5 Content MD5 specified for the blob.
-     * @param contentEncoding Content encoding specified for the blob.
-     * @param contentDisposition Content disposition specified for the blob.
-     * @param contentLanguage Content language specified for the blob.
-     * @param cacheControl Cache control specified for the blob.
-     * @param blobSequenceNumber The current sequence number for a page blob, if the blob is an append or block blob
-     * pass {@code null}.
-     * @param blobType Type of the blob.
-     * @param leaseStatus Status of the lease on the blob.
-     * @param leaseState State of the lease on the blob.
-     * @param leaseDuration Type of lease on the blob.
-     * @param copyId Identifier of the last copy operation performed on the blob.
-     * @param copyStatus Status of the last copy operation performed on the blob.
-     * @param copySource Source of the last copy operation performed on the blob.
-     * @param copyProgress Progress of the last copy operation performed on the blob.
-     * @param copyCompletionTime Datetime when the last copy operation on the blob completed.
-     * @param copyStatusDescription Description of the last copy operation on the blob.
-     * @param isServerEncrypted Flag indicating if the blob's content is encrypted on the server.
-     * @param isIncrementalCopy Flag indicating if the blob was incrementally copied.
-     * @param copyDestinationSnapshot Snapshot identifier of the last incremental copy snapshot for the blob.
-     * @param accessTier Access tier of the blob.
-     * @param isAccessTierInferred Flag indicating if the access tier of the blob was inferred from properties of the
-     * blob.
-     * @param archiveStatus Archive status of the blob.
-     * @param encryptionKeySha256 SHA256 of the customer provided encryption key used to encrypt the blob on the server.
-     * @param accessTierChangeTime Datetime when the access tier of the blob last changed.
-     * @param metadata Metadata associated with the blob.
-     * @param committedBlockCount Number of blocks committed to an append blob, if the blob is a block or page blob
-     * pass {@code null}.
-     * @param versionId The version identifier of the blob.
-     * @param isCurrentVersion Flag indicating if version identifier points to current version of the blob.
+     * @param tagCount Number of tags associated with the blob.
      * @param objectReplicationStatus The object replication status map to parse.
      * @param expiresOn Datetime when the blob is going to expire.
      */
@@ -220,10 +164,9 @@ public final class BlobProperties {
         final OffsetDateTime copyCompletionTime, final String copyStatusDescription, final Boolean isServerEncrypted,
         final Boolean isIncrementalCopy, final String copyDestinationSnapshot, final AccessTier accessTier,
         final Boolean isAccessTierInferred, final ArchiveStatus archiveStatus, final String encryptionKeySha256,
-        final OffsetDateTime accessTierChangeTime, final Map<String, String> metadata,
+        final String encryptionScope, final OffsetDateTime accessTierChangeTime, final Map<String, String> metadata,
         final Integer committedBlockCount, final String versionId, final Boolean isCurrentVersion,
-        Map<String, String> objectReplicationStatus, final OffsetDateTime expiresOn) {
-
+        final Long tagCount, Map<String, String> objectReplicationStatus, final OffsetDateTime expiresOn) {
         this.creationTime = creationTime;
         this.lastModified = lastModified;
         this.eTag = eTag;
@@ -252,9 +195,11 @@ public final class BlobProperties {
         this.isAccessTierInferred = isAccessTierInferred;
         this.archiveStatus = archiveStatus;
         this.encryptionKeySha256 = encryptionKeySha256;
+        this.encryptionScope = encryptionScope;
         this.accessTierChangeTime = accessTierChangeTime;
         this.metadata = metadata;
         this.committedBlockCount = committedBlockCount;
+        this.tagCount = tagCount;
         this.versionId = versionId;
         this.isCurrentVersion = isCurrentVersion;
 
@@ -312,12 +257,14 @@ public final class BlobProperties {
      * blob.
      * @param archiveStatus Archive status of the blob.
      * @param encryptionKeySha256 SHA256 of the customer provided encryption key used to encrypt the blob on the server.
+     * @param encryptionScope The name of the encryption scope under which the blob is encrypted.
      * @param accessTierChangeTime Datetime when the access tier of the blob last changed.
      * @param metadata Metadata associated with the blob.
      * @param committedBlockCount Number of blocks committed to an append blob, if the blob is a block or page blob
      * pass {@code null}.
      * @param versionId The version identifier of the blob.
      * @param isCurrentVersion Flag indicating if version identifier points to current version of the blob.
+     * @param tagCount Number of tags associated with the blob.
      * @param objectReplicationStatus The already parsed object replication status.
      * @param objectReplicationDestinationPolicyId The policy id on the destination blob.
      */
@@ -330,9 +277,10 @@ public final class BlobProperties {
         final OffsetDateTime copyCompletionTime, final String copyStatusDescription, final Boolean isServerEncrypted,
         final Boolean isIncrementalCopy, final String copyDestinationSnapshot, final AccessTier accessTier,
         final Boolean isAccessTierInferred, final ArchiveStatus archiveStatus, final String encryptionKeySha256,
-        final OffsetDateTime accessTierChangeTime, final Map<String, String> metadata,
-        final Integer committedBlockCount, final String versionId, final Boolean isCurrentVersion,
-        Map<String, ObjectReplicationPolicy> objectReplicationStatus, String objectReplicationDestinationPolicyId) {
+        String encryptionScope, final OffsetDateTime accessTierChangeTime, final Map<String, String> metadata,
+        final Integer committedBlockCount, final Long tagCount, final String versionId,
+        final Boolean isCurrentVersion, Map<String, ObjectReplicationPolicy> objectReplicationStatus,
+        String objectReplicationDestinationPolicyId) {
 
         this.creationTime = creationTime;
         this.lastModified = lastModified;
@@ -362,9 +310,11 @@ public final class BlobProperties {
         this.isAccessTierInferred = isAccessTierInferred;
         this.archiveStatus = archiveStatus;
         this.encryptionKeySha256 = encryptionKeySha256;
+        this.encryptionScope = encryptionScope;
         this.accessTierChangeTime = accessTierChangeTime;
         this.metadata = metadata;
         this.committedBlockCount = committedBlockCount;
+        this.tagCount = tagCount;
         this.versionId = versionId;
         this.isCurrentVersion = isCurrentVersion;
         this.objectReplicationSourcePolicies = objectReplicationStatus;
@@ -580,6 +530,13 @@ public final class BlobProperties {
     }
 
     /**
+     * @return The name of the encryption scope under which the blob is encrypted.
+     */
+    public String getEncryptionScope() {
+        return encryptionScope;
+    }
+
+    /**
      * @return the time when the access tier for the blob was last changed
      */
     public OffsetDateTime getAccessTierChangeTime() {
@@ -598,6 +555,13 @@ public final class BlobProperties {
      */
     public Integer getCommittedBlockCount() {
         return committedBlockCount;
+    }
+
+    /**
+     * @return The number of tags associated with the blob.
+     */
+    public Long getTagCount() {
+        return tagCount;
     }
 
     /**