+++ sdk/storage/azure-storage-blob/src/main/java/com/azure/storage/blob/models/BlockBlobSimpleUploadOptions.java
@@ -1,0 +1,122 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.storage.blob.models;
+
+import com.azure.core.util.CoreUtils;
+
+import java.util.Map;
+
+/**
+ * Extended options that may be passed when uploading a Block Blob in a single request.
+ */
+public class BlockBlobSimpleUploadOptions {
+    private BlobHttpHeaders headers;
+    private Map<String, String> metadata;
+    private Map<String, String> tags;
+    private AccessTier tier;
+    private byte[] contentMd5;
+    private BlobRequestConditions requestConditions;
+
+    /**
+     * @return {@link BlobHttpHeaders}
+     */
+    public BlobHttpHeaders getHeaders() {
+        return headers;
+    }
+
+    /**
+     * @param headers {@link BlobHttpHeaders}
+     * @return The updated {@code AppendBlobCreateOptions}
+     */
+    public BlockBlobSimpleUploadOptions setHeaders(BlobHttpHeaders headers) {
+        this.headers = headers;
+        return this;
+    }
+
+    /**
+     * @return The metadata to associate with the blob.
+     */
+    public Map<String, String> getMetadata() {
+        return metadata;
+    }
+
+    /**
+     * @param metadata The metadata to associate with the blob.
+     * @return The updated options
+     */
+    public BlockBlobSimpleUploadOptions setMetadata(Map<String, String> metadata) {
+        this.metadata = metadata;
+        return this;
+    }
+
+    /**
+     * @return The tags to associate with the blob.
+     */
+    public Map<String, String> getTags() {
+        return tags;
+    }
+
+    /**
+     * @param tags The tags to associate with the blob.
+     * @return The updated options.
+     */
+    public BlockBlobSimpleUploadOptions setTags(Map<String, String> tags) {
+        this.tags = tags;
+        return this;
+    }
+
+    /**
+     * @return {@link AccessTier}
+     */
+    public AccessTier getTier() {
+        return tier;
+    }
+
+    /**
+     * @param tier {@link AccessTier}
+     * @return The updated options.
+     */
+    public BlockBlobSimpleUploadOptions setTier(AccessTier tier) {
+        this.tier = tier;
+        return this;
+    }
+
+    /**
+     * @return An MD5 hash of the block content. This hash is used to verify the integrity of the block during
+     * transport. When this header is specified, the storage service compares the hash of the content that has arrived
+     * with this header value. Note that this MD5 hash is not stored with the blob. If the two hashes do not match, the
+     * operation will fail.
+     */
+    public byte[] getContentMd5() {
+        return CoreUtils.clone(contentMd5);
+    }
+
+    /**
+     * @param contentMd5 An MD5 hash of the block content. This hash is used to verify the integrity of the block during
+     * transport. When this header is specified, the storage service compares the hash of the content that has arrived
+     * with this header value. Note that this MD5 hash is not stored with the blob. If the two hashes do not match, the
+     * operation will fail.
+     * @return The updated options
+     */
+    public BlockBlobSimpleUploadOptions setContentMd5(byte[] contentMd5) {
+        this.contentMd5 = CoreUtils.clone(contentMd5);
+        return this;
+    }
+
+    /**
+     * @return {@link BlobRequestConditions}
+     */
+    public BlobRequestConditions getRequestConditions() {
+        return requestConditions;
+    }
+
+    /**
+     * @param requestConditions {@link BlobRequestConditions}
+     * @return The updated options.
+     */
+    public BlockBlobSimpleUploadOptions setRequestConditions(BlobRequestConditions requestConditions) {
+        this.requestConditions = requestConditions;
+        return this;
+    }
+}